+++ lib/src/wheel_picker_controller.dart
@@ -10,6 +10,7 @@ part './wheel_picker.dart';
 /// The `WheelPickerController` allows you to:
 /// - Retrieve the selected item index.
 /// - Shift the attached `WheelPicker` widget.
+/// - Set the current item selected.
 /// - Mount additional `WheelPickerController` instances, which will be shifted when looped.
 ///
 /// It provides precise control over scrolling and synchronization of multiple `WheelPicker` widgets.
@@ -19,8 +20,10 @@ part './wheel_picker.dart';
 /// - Connect it to one or more `WheelPicker` widgets using the `controller` property.
 /// - Use its methods to manipulate the selected item index and scrolling behavior.
 class WheelPickerController {
-  /// The total number of items in the wheel.
-  final int itemCount;
+  static const _defaultInitialIndex = 0;
+
+  /// The total number of items in the wheel - (can be changed).
+  int itemCount;
 
   /// The initial selected item index.
   final int initialIndex;
@@ -54,12 +57,15 @@ class WheelPickerController {
   /// Note: Make sure to manually dispose of the controller when it is no longer needed to release associated resources.
   WheelPickerController({
     required this.itemCount,
-    this.initialIndex = 0,
+    this.initialIndex = WheelPickerController._defaultInitialIndex,
     List<WheelPickerController>? mounts,
   })  : _mounts = mounts ?? [],
         _scrollController =
             FixedExtentScrollController(initialItem: initialIndex),
         _current = initialIndex;
+
+  /// Caps index to be ranging from 0 to `itemCount` - 1
+  int _capIndex(int index) => math.min(math.max(index, 0), itemCount - 1);
 
   /// Wether scroll controller is attached to a mounted widget.
   bool get _hasClients => _scrollController.hasClients;
@@ -149,6 +155,29 @@ class WheelPickerController {
     return await _shiftController(this, steps, _shiftAnimationStyle);
   }
 
+  /// Shifts the wheel to the specified index.
+  ///
+  /// Parameters:
+  /// - `index`: The item number to jump to ranging from 0 to `itemCount` - 1.
+  Future<void> shiftTo({required int index}) async {
+    if (!_hasClients) return;
+    return await _shiftController(
+      this,
+      _capIndex(index),
+      _shiftAnimationStyle,
+      relativeToCurrent: false,
+    );
+  }
+
+  /// Sets the current item to the specified index.
+  ///
+  /// Parameters:
+  /// - `index`: The item number to jump to ranging from 0 to `itemCount` - 1
+  Future<void> setCurrent(int index) async {
+    if (!_hasClients) return;
+    _jumpController(this, _capIndex(index));
+  }
+
   /// The current selected item index.
   ///
   /// Returns:
@@ -175,12 +204,29 @@ class WheelPickerController {
 Future<void> _shiftController(
   WheelPickerController controller,
   int steps,
-  WheelShiftAnimationStyle shiftAnimationStyle,
-) async {
+  WheelShiftAnimationStyle shiftAnimationStyle, {
+  bool relativeToCurrent = true,
+}) async {
   if (!controller._hasClients) return;
   return await controller._scrollController.animateToItem(
-    controller._scrollController.selectedItem + steps,
+    (relativeToCurrent ? controller._scrollController.selectedItem : 0) + steps,
     duration: shiftAnimationStyle.duration,
     curve: shiftAnimationStyle.curve,
   );
+}
+
+/// Jumps the controller's scroll position to the item at the given index.
+/// Parameters:
+/// - `controller`: The controller managing the scroll.
+/// - `index`: The item number to jump to ranging from 0 to `itemCount` - 1.
+///
+/// This method ensures the controller has clients before jumping.
+///
+/// * Note that this is only supposed to be in the controller scope itself.
+Future<void> _jumpController(
+  WheelPickerController controller,
+  int index,
+) async {
+  if (!controller._hasClients) return;
+  controller._scrollController.jumpToItem(index);
 }